chore(mm-routing): drop unnecessary self->usize cast
self.kv_cache_block_size is already usize since the duplicate u32 field
was removed; clippy::unnecessary_cast flags the now-redundant `as usize`.

+++ lib/llm/src/preprocessor.rs
@@ -847,7 +847,7 @@ impl OpenAIPreprocessor {
             );
             return Ok(());
         };
-        let block_size = self.kv_cache_block_size as usize;
+        let block_size = self.kv_cache_block_size;
         if block_size == 0 {
             tracing::debug!(
                 target: "mm_routing",
